refactoring Categories and Charts code
